// src/generated/models/Coils_CollarsBuyoffStructureModel.ts
export interface Machine_x0023_Value {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Machine_x0020__x0023__x003a_AreaValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface MaterialValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface MaterialTypeValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface OData__x0023_ofCavsValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Deviation_x003f__x0028_Visual_x0Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DeviationPending_x003f__x0028_ViValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface Deviation_x003f__x0028_DimensionValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DeviationPending_x003f__x0028_DiValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyoffAccepted_x0028_Visual_x002Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyoffAccepted_x0028_DimensionalValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface ShiftValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface PlantValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface NTQValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface BuyOffAccepted_x003f_Value {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface DefectDescriptionValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface Defect_x0020_Description_x003a_DValue {
  "@odata.type": string;
  Id: number;
  Value: string;
}

export interface BuyoffTypeValue {
  "@odata.type": string;
  Value: string;
  Id: number;
}

export interface AuthorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface EditorValue {
  "@odata.type": string;
  Claims: string;
  DisplayName: string;
  Email: string;
  Picture: string;
  Department: string;
  JobTitle: string;
}

export interface Coils_CollarsBuyoffStructureBase {
  ID?: number;
  Title?: string;
  DateCreated?: string;
  TimeIn?: string;
  WorkOrder_x0023_?: string;
  "Machine_x0023_#Id"?: number;
  "Machine_x0020__x0023__x003a_Area#Id"?: number;
  Part_x0023_?: string;
  PartDescription?: string;
  "Material#Id"?: number;
  "MaterialType#Id"?: number;
  OtherMaterial?: string;
  "OData__x0023_ofCavs#Id"?: number;
  MoldingTech?: string;
  Comments?: string;
  "Deviation_x003f__x0028_Visual_x0#Id"?: number[];
  "Deviation_x003f__x0028_Visual_x0@odata.type"?: string;
  "DeviationPending_x003f__x0028_Vi#Id"?: number;
  DeviationNumber_x0028_Visual_x00?: string;
  DeviationDescription_x0028_Visua?: string;
  "Deviation_x003f__x0028_Dimension#Id"?: number[];
  "Deviation_x003f__x0028_Dimension@odata.type"?: string;
  "DeviationPending_x003f__x0028_Di#Id"?: number;
  DeviationNumber_x0028_Dimensiona?: string;
  DeviationDescription_x0028_Dimen?: string;
  "BuyoffAccepted_x0028_Visual_x002#Id"?: number;
  QCTech_x0028_Visual_x0029_?: string;
  TimeOut_x0028_Visual_x0029_?: string;
  "BuyoffAccepted_x0028_Dimensional#Id"?: number;
  QCTech_x0028_Dimensional_x0029_?: string;
  TimeOut_x0028_Dimensional_x0029_?: string;
  "Shift#Id"?: number;
  "Plant#Id"?: number;
  QCComments?: string;
  Time_x0020_Open?: string;
  Overall_x0020_Acceptance?: string;
  OData__ColorTag?: string;
  ComplianceAssetId?: string;
  "NTQ#Id"?: number;
  SprueWeight?: string;
  PartWeight?: string;
  Attach?: string;
  TimeOut?: string;
  QCTech?: string;
  QtyScrapped?: string;
  "BuyOffAccepted_x003f_#Id"?: number;
  "DefectDescription#Id"?: number[];
  "DefectDescription@odata.type"?: string;
  "Defect_x0020_Description_x003a_D#Id"?: number[];
  "Defect_x0020_Description_x003a_D@odata.type"?: string;
  Date_x0020_Completed_x0020__x002?: string;
  Date_x0020_Completed_x0020__x0020?: string;
  Consolidated_x0020_Deviation_x00?: string;
  "BuyoffType#Id"?: number;
  Modified?: string;
  Created?: string;
  "Author#Claims"?: string;
  "Editor#Claims"?: string;
  "{Identifier}"?: string;
  "{IsFolder}"?: boolean;
  "{Thumbnail}"?: {
    "@odata.type"?: string;
    // URL to the small, highly compressed size thumbnail of the item, if available.
    Small?: string;
    // URL to the standard size thumbnail of the item, if available.
    Medium?: string;
    // URL to the largest size thumbnail of the item, if available.
    Large?: string;
  };
  "{Link}"?: string;
  "{Name}"?: string;
  "{FilenameWithExtension}"?: string;
  "{Path}"?: string;
  "{FullPath}"?: string;
  "{ModerationStatus}"?: string;
  "{ModerationComment}"?: string;
  "{ContentType}#Id"?: string;
  "{ContentType}"?: {
    "@odata.type"?: string;
    Id?: string;
    Name?: string;
  };
  "{HasAttachments}"?: boolean;
  "{Attachments}@odata.type"?: string;
  "{Attachments}"?: {
   "@odata.type"?: string;
   Id?: string;
   AbsoluteUri?: string;
   DisplayName?: string;
 }[];
  "{VersionNumber}"?: string;
  "{TriggerWindowStartToken}"?: string;
  "{TriggerWindowEndToken}"?: string;
}

export interface Coils_CollarsBuyoffStructureRead extends Coils_CollarsBuyoffStructureBase {
  Machine_x0023_?: Machine_x0023_Value;
  Machine_x0020__x0023__x003a_Area?: Machine_x0020__x0023__x003a_AreaValue;
  Material?: MaterialValue;
  MaterialType?: MaterialTypeValue;
  OData__x0023_ofCavs?: OData__x0023_ofCavsValue;
  Deviation_x003f__x0028_Visual_x0?: Deviation_x003f__x0028_Visual_x0Value;
  DeviationPending_x003f__x0028_Vi?: DeviationPending_x003f__x0028_ViValue;
  Deviation_x003f__x0028_Dimension?: Deviation_x003f__x0028_DimensionValue;
  DeviationPending_x003f__x0028_Di?: DeviationPending_x003f__x0028_DiValue;
  BuyoffAccepted_x0028_Visual_x002?: BuyoffAccepted_x0028_Visual_x002Value;
  BuyoffAccepted_x0028_Dimensional?: BuyoffAccepted_x0028_DimensionalValue;
  Shift?: ShiftValue;
  Plant?: PlantValue;
  NTQ?: NTQValue;
  BuyOffAccepted_x003f_?: BuyOffAccepted_x003f_Value;
  DefectDescription?: DefectDescriptionValue;
  Defect_x0020_Description_x003a_D?: Defect_x0020_Description_x003a_DValue;
  BuyoffType?: BuyoffTypeValue;
  Author?: AuthorValue;
  Editor?: EditorValue;
}